+++ doc/sphinx/changelog.rst
@@ -9,6 +9,8 @@ New
 
 - Implement user-defined literals for :cpp:class:`~mppp::integer`
   (`#209 <https://github.com/bluescarni/mppp/pull/209>`__).
+- Implement (modular) squaring primitives for :cpp:class:`~mppp::integer`
+  (`#205 <https://github.com/bluescarni/mppp/pull/205>`__).
 - Include mp++'s headers in the project files generated
   for MSVC (`#199 <https://github.com/bluescarni/mppp/pull/199>`__).
   Many thanks to `7ofNine <https://github.com/7ofNine>`__!